pyproject.toml
[project]
name = "tasklist_agent_crew_ai"
version = "0.1.0"
description = "tasklist_agent_crew_ai using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "python-jose>=3.4.0",
]

[project.scripts]
tasklist_agent_crew_ai = "tasklist_agent_crew_ai.main:run"
run_crew = "tasklist_agent_crew_ai.main:run"
train = "tasklist_agent_crew_ai.main:train"
replay = "tasklist_agent_crew_ai.main:replay"
test = "tasklist_agent_crew_ai.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
